unnecessary code removal

+++ player/player.py
@@ -40,10 +40,10 @@ class Player:
           
      def skip(self,time_in_seconds: int):
           current_time = self.player.get_time()
-          if time_in_seconds + current_time/1000 >= self.player.get_length()/1000:
+          if time_in_seconds + int(current_time/1000) >= self.player.get_length()/1000:
                print("Skipping beyond the end of the media is not allowed.")
                return
-          if time_in_seconds + current_time/1000 < 0:
+          if time_in_seconds + int(current_time/1000) < 0:
                print("Skipping to a negative time is not allowed.")
                return
           self.player.pause()
